+++ kernel/mpdb-queue.h
@@ -33,7 +33,7 @@ extern bool
 mpdb_enqueue(struct pfq_opt *pq, struct sk_buff *skb);
 
 extern void *
-mpdb_queue_alloc(struct pfq_opt *pq, int queue_mem, size_t * tot_mem);
+mpdb_queue_alloc(struct pfq_opt *pq, size_t queue_mem, size_t * tot_mem);
 
 extern void
 mpdb_queue_free(struct pfq_opt *pq);
@@ -59,7 +59,7 @@ static inline
 size_t
 mpdb_queue_size(struct pfq_opt *pq)
 {
-        return sizeof(struct pfq_queue_descr) + pq->q_slot_size * pq->q_slots * 2; 
+    return sizeof(struct pfq_queue_descr) + pq->q_slot_size * pq->q_slots * 2; 
 }
 
 #endif /* _MPDB_QUEUE_H_ */
